dropdown now works, added folders for pages and adjusted color so it should be green when on light mode
