+++ shelfShare/app/books/page.tsx
@@ -1,0 +1,12 @@
+// pages/books.tsx
+import React from 'react'
+
+const Books = () => {
+    return (
+        <div className="container mt-2">
+            This is a books page
+        </div>
+    )
+}
+
+export default Books
