+++ shelfShare/app/browse/page.tsx
@@ -1,0 +1,12 @@
+// pages/browse.tsx
+import React from 'react'
+
+const Browse = () => {
+    return (
+        <div className="container mt-2">
+            This is a browse page
+        </div>
+    )
+}
+
+export default Browse
