+++ shelfShare/app/friends/page.tsx
@@ -1,0 +1,12 @@
+// pages/friends.tsx
+import React from 'react'
+
+const Friends = () => {
+    return (
+        <div className="container mt-2">
+            This is a friends page
+        </div>
+    )
+}
+
+export default Friends
