+++ shelfShare/app/globals.css
@@ -4,7 +4,7 @@
 @tailwind utilities;
 
 :root {
-  --background: #ffffff;
+  --background: #5a7463;
   --foreground: #171717;
 }
