+++ shelfShare/app/logout/page.tsx
@@ -1,0 +1,12 @@
+// pages/logout.tsx
+import React from 'react'
+
+const Logout = () => {
+    return (
+        <div className="container mt-2">
+            This is a logout page
+        </div>
+    )
+}
+
+export default Logout
