+++ shelfShare/app/movies/page.tsx
@@ -1,0 +1,12 @@
+// pages/movies.tsx
+import React from 'react'
+
+const Movies = () => {
+    return (
+        <div className="container mt-2">
+            This is a movies page
+        </div>
+    )
+}
+
+export default Movies
