+++ shelfShare/app/music/page.tsx
@@ -1,14 +1,11 @@
 // pages/music.tsx
-import Layout from '@/app/layout'
 import React from 'react'
 
 const Music = () => {
     return (
-        <Layout >
             <div className="container mt-2">
                 This is a music page
             </div>
-        </Layout>
     )
 }
