+++ shelfShare/app/my-shelf/page.tsx
@@ -1,0 +1,12 @@
+// pages/my-shelf.tsx
+import React from 'react'
+
+const myShelf = () => {
+    return (
+        <div className="container mt-2">
+            This is a my shelf page
+        </div>
+    )
+}
+
+export default myShelf
