+++ shelfShare/app/profile/page.tsx
@@ -1,0 +1,12 @@
+// pages/profile.tsx
+import React from 'react'
+
+const Profile = () => {
+    return (
+        <div className="container mt-2">
+            This is a profile page
+        </div>
+    )
+}
+
+export default Profile
