+++ shelfShare/app/settings/page.tsx
@@ -1,0 +1,12 @@
+// pages/settings.tsx
+import React from 'react'
+
+const Settings = () => {
+    return (
+        <div className="container mt-2">
+            This is a settings page
+        </div>
+    )
+}
+
+export default Settings
