+++ shelfShare/app/timeline/page.tsx
@@ -1,0 +1,12 @@
+// pages/timeline.tsx
+import React from 'react'
+
+const Timeline = () => {
+    return (
+        <div className="container mt-2">
+            This is a timeline page
+        </div>
+    )
+}
+
+export default Timeline
