--- shelfShare/pages/books.tsx
@@ -1,15 +1,0 @@
-// pages/books.tsx
-import Layout from '@/app/layout'
-import React from 'react'
-
-const Books = () => {
-    return (
-        <Layout >
-            <div className="container mt-2">
-                This is a books page
-            </div>
-        </Layout>
-    )
-}
-
-export default Books
