--- shelfShare/pages/browse.tsx
@@ -1,15 +1,0 @@
-// pages/browse.tsx
-import Layout from '@/app/layout'
-import React from 'react'
-
-const Browse = () => {
-    return (
-        <Layout >
-            <div className="container mt-2">
-                This is a browse page
-            </div>
-        </Layout>
-    )
-}
-
-export default Browse
