--- shelfShare/pages/friends.tsx
@@ -1,15 +1,0 @@
-// pages/friends.tsx
-import Layout from '@/app/layout'
-import React from 'react'
-
-const Friends = () => {
-    return (
-        <Layout >
-            <div className="container mt-2">
-                This is a friends page
-            </div>
-        </Layout>
-    )
-}
-
-export default Friends
