--- shelfShare/pages/logout.tsx
@@ -1,15 +1,0 @@
-// pages/logout.tsx
-import Layout from '@/app/layout'
-import React from 'react'
-
-const Logout = () => {
-    return (
-        <Layout >
-            <div className="container mt-2">
-                This is a logout page
-            </div>
-        </Layout>
-    )
-}
-
-export default Logout
