--- shelfShare/pages/movies.tsx
@@ -1,15 +1,0 @@
-// pages/movies.tsx
-import Layout from '@/app/layout'
-import React from 'react'
-
-const Movies = () => {
-    return (
-        <Layout >
-            <div className="container mt-2">
-                This is a movies page
-            </div>
-        </Layout>
-    )
-}
-
-export default Movies
