--- shelfShare/pages/my-shelf.tsx
@@ -1,15 +1,0 @@
-// pages/my-shelf.tsx
-import Layout from '@/app/layout'
-import React from 'react'
-
-const myShelf = () => {
-    return (
-        <Layout >
-            <div className="container mt-2">
-                This is a my shelf page
-            </div>
-        </Layout>
-    )
-}
-
-export default myShelf
