--- shelfShare/pages/profile.tsx
@@ -1,15 +1,0 @@
-// pages/profile.tsx
-import Layout from '@/app/layout'
-import React from 'react'
-
-const Profile = () => {
-    return (
-        <Layout >
-            <div className="container mt-2">
-                This is a profile page
-            </div>
-        </Layout>
-    )
-}
-
-export default Profile
