--- shelfShare/pages/settings.tsx
@@ -1,15 +1,0 @@
-// pages/settings.tsx
-import Layout from '@/app/layout'
-import React from 'react'
-
-const Settings = () => {
-    return (
-        <Layout >
-            <div className="container mt-2">
-                This is a settings page
-            </div>
-        </Layout>
-    )
-}
-
-export default Settings
