--- shelfShare/pages/timeline.tsx
@@ -1,15 +1,0 @@
-// pages/timeline.tsx
-import Layout from '@/app/layout'
-import React from 'react'
-
-const Timeline = () => {
-    return (
-        <Layout >
-            <div className="container mt-2">
-                This is a timeline page
-            </div>
-        </Layout>
-    )
-}
-
-export default Timeline
